src/services/webhook-listener/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
helius==0.0.3
httpx==0.25.2
python-dotenv==1.0.0
# Database dependencies
sqlalchemy==2.0.23
sqlmodel==0.0.14
psycopg2-binary==2.9.9
# Remove Delta Lake dependencies
# duckdb==1.3.0
# pandas==2.2.0
# pyarrow==15.0.0
# s3fs==2024.2.0
# deltalake==0.15.0